+++ src/features/candidate/Profile/components/CandidateProfile.tsx
@@ -4,9 +4,7 @@ import { useEffect, useState } from "react";
 import { useSelector } from "react-redux";
 import { RootState } from "@/config/store";
 
-import { useGetProfile, useCreateProfile, useUpdateMe } from "../hooks/useProfileApi";
-import Modal from "@/components/ui/Modal";
-import { LoaderCircleIcon } from "lucide-react";
+import { useGetProfile, useCreateProfile } from "../hooks/useProfileApi";
 
 import PersonalInfoSection from "./PersonalInfoSection";
 import SkillsSection from "./SkillsSection";
@@ -15,6 +13,7 @@ import ResumeSection from "./ResumeSection";
 import SocialLinksSection from "./SocialLinksSection";
 import AvailabilitySection from "./AvailabilitySection";
 import ProfileCompletion from "./ProfileCompletion";
+import EditProfileInfoModal from "./EditProfileInfoModal";
 
 export default function CandidateProfile() {
   const authUser = useSelector((state: RootState) => state.auth.user);
@@ -70,7 +69,7 @@ export default function CandidateProfile() {
             phone={profile?.user?.phoneNumber ?? ""}
           />
 
-          <EditPersonalInfoModal
+          <EditProfileInfoModal
             profile={profile}
             isOpen={isEditOpen}
             onClose={toggleEdit}
@@ -137,111 +136,5 @@ export default function CandidateProfile() {
         </div>
       </div>
     </div>
-  );
-}
-
-// @ts-expect-error -- props are untyped here
-function EditPersonalInfoModal(props) {
-  const { profile, isOpen, onClose, onUpdated } = props;
-
-  const [firstName, setFirstName] = useState(profile?.user?.firstName ?? "");
-  const [lastName, setLastName] = useState(profile?.user?.lastName ?? "");
-  const [phone, setPhone] = useState(profile?.user?.phoneNumber ?? "");
-
-  const { mutate: updateMe, isPending } = useUpdateMe();
-
-  // sync when modal opens or profile changes
-  useEffect(() => {
-    if (isOpen) {
-      setFirstName(profile?.user?.firstName ?? "");
-      setLastName(profile?.user?.lastName ?? "");
-      setPhone(profile?.user?.phoneNumber ?? "");
-    }
-  }, [isOpen, profile?.user?.firstName, profile?.user?.lastName, profile?.user?.phoneNumber]);
-
-  const handleSave = () => {
-    // JWT handles identity, no need for userId
-    updateMe(
-      {
-        firstName: firstName.trim(),
-        lastName: lastName.trim(),
-        phoneNumber: phone.trim(),
-      },
-      {
-        onSuccess: () => {
-          onClose();
-          onUpdated?.();
-        },
-        onError: (err) => {
-          console.error("Update error:", err);
-          alert("Failed to update personal info. Please try again.");
-        },
-      }
-    );
-  };
-
-  return (
-    <Modal isOpen={isOpen} onClose={onClose} title="Edit Personal Information">
-      <div className="space-y-4">
-        <div>
-          <label className="block text-sm text-gray-700 mb-1">First Name</label>
-          <input
-            value={firstName}
-            onChange={(e) => setFirstName(e.target.value)}
-            className="w-full px-4 py-2 border border-gray-300 rounded-lg"
-          />
-        </div>
-
-        <div>
-          <label className="block text-sm text-gray-700 mb-1">Last Name</label>
-          <input
-            value={lastName}
-            onChange={(e) => setLastName(e.target.value)}
-            className="w-full px-4 py-2 border border-gray-300 rounded-lg"
-          />
-        </div>
-
-        <div>
-          <label className="block text-sm text-gray-700 mb-1">Email (readonly)</label>
-          <input
-            value={profile?.user?.email ?? ""}
-            readOnly
-            className="w-full px-4 py-2 border border-gray-200 bg-gray-100 rounded-lg"
-          />
-        </div>
-
-        <div>
-          <label className="block text-sm text-gray-700 mb-1">Phone</label>
-          <input
-            value={phone}
-            onChange={(e) => setPhone(e.target.value)}
-            className="w-full px-4 py-2 border border-gray-300 rounded-lg"
-          />
-        </div>
-
-        <div className="flex justify-end gap-3 pt-4">
-          <button
-            onClick={onClose}
-            className="px-5 py-2 border border-gray-300 rounded-lg hover:bg-gray-50"
-          >
-            Cancel
-          </button>
-          <button
-            onClick={handleSave}
-            disabled={isPending}
-            className="px-6 py-2 bg-blue-600 text-white rounded-lg hover:bg-blue-700 disabled:opacity-50 flex items-center gap-2"
-          >
-            {isPending ? (
-              <>
-                <LoaderCircleIcon className="animate-spin w-5 h-5" />
-                Saving...
-              </>
-            ) : (
-              "Save"
-            )}
-          </button>
-        </div>
-      </div>
-    </Modal>
   );
 }
